mcp-server/requirements.txt
fastapi
uvicorn
pydantic
pygithub
gitpython
openai
redis
celery
markdown
black
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
aiofiles
jinja2
pyyaml